@@ -15,29 +15,28 @@
         "source": "https://github.com/Mapiiik/Watcher-NMS"
     },
     "require": {
-        "php": ">=7.4",
-        "cakedc/users": "^11.2",
-        "cakephp/cakephp": "^4.4",
-        "cakephp/migrations": "^3.2",
-        "cakephp/plugin-installer": "^1.3",
-        "dereuromark/cakephp-geo": "^2.2",
+        "php": ">=8.1",
+        "cakedc/users": "dev-12.next-cake5",
+        "cakephp/cakephp": "^5.0",
+        "cakephp/migrations": "4.x-dev",
+        "cakephp/plugin-installer": "^2.0",
+        "dereuromark/cakephp-geo": "dev-cake5",
         "geocoder-php/google-maps-provider": "^4.4",
         "league/oauth2-google": "^4.0",
-        "mapik/audit-log": "^1.0",
+        "mapik/audit-log": "dev-master",
         "mobiledetect/mobiledetectlib": "^3.74",
-        "php-http/cakephp-adapter": "^0.3.0",
+        "php-http/cakephp-adapter": "dev-next",
         "php-http/message": "^1.8.0"
     },
     "require-dev": {
-        "cakedc/cakephp-phpstan": "^2.0",
-        "cakephp/bake": "^2.6",
-        "cakephp/cakephp-codesniffer": "^4.5",
-        "cakephp/debug_kit": "^4.5",
-        "cakephp/repl": "^0.1",
-        "josegonzalez/dotenv": "^3.2",
+        "cakedc/cakephp-phpstan": "dev-3.next-cake5",
+        "cakephp/bake": "3.x-dev",
+        "cakephp/cakephp-codesniffer": "^5.0",
+        "cakephp/debug_kit": "5.x-dev",
+        "josegonzalez/dotenv": "^4.0",
         "phpstan/extension-installer": "^1.1",
         "phpstan/phpstan": "^1.0",
-        "phpunit/phpunit": "~8.5.0 || ^9.3",
+        "phpunit/phpunit": "^10.1.0",
         "psalm/phar": "^5.12"
     },
     "suggest": {
@@ -82,6 +81,7 @@
         "test": "phpunit --colors=always"
     },
     "prefer-stable": true,
+    "minimum-stability": "dev",
     "config": {
         "sort-packages": true,
         "allow-plugins": {
