@@ -17,7 +17,7 @@ class StringModificationsBehavior extends Behavior
      *
      * @var array<string, mixed>
      */
-    protected $_defaultConfig = [
+    protected array $_defaultConfig = [
         'trim' => true,
         'emptyAsNull' => true,
         'replaceBadCharacters' => true,
