@@ -13,10 +13,10 @@ use Cake\ORM\Entity;
  * @property string|null $identifier
  * @property string|null $snmp_community
  * @property string|null $note
- * @property \Cake\I18n\FrozenTime|null $created
+ * @property \Cake\I18n\DateTime|null $created
  * @property string|null $created_by
  * @property \CakeDC\Users\Model\Entity\User|null $creator
- * @property \Cake\I18n\FrozenTime|null $modified
+ * @property \Cake\I18n\DateTime|null $modified
  * @property string|null $modified_by
  * @property \CakeDC\Users\Model\Entity\User|null $modifier
  * @property bool $assign_access_point_by_device_name
@@ -39,7 +39,7 @@ class DeviceType extends Entity
      *
      * @var array<string, bool>
      */
-    protected $_accessible = [
+    protected array $_accessible = [
         'name' => true,
         'identifier' => true,
         'snmp_community' => true,
