@@ -11,13 +11,13 @@ use Cake\ORM\Entity;
  * @property string $id
  * @property string|null $name
  * @property string|null $access_point_id
- * @property \Cake\I18n\FrozenDate|null $reading_date
+ * @property \Cake\I18n\Date|null $reading_date
  * @property float|null $reading_value
  * @property string|null $note
- * @property \Cake\I18n\FrozenTime|null $created
+ * @property \Cake\I18n\DateTime|null $created
  * @property string|null $created_by
  * @property \CakeDC\Users\Model\Entity\User|null $creator
- * @property \Cake\I18n\FrozenTime|null $modified
+ * @property \Cake\I18n\DateTime|null $modified
  * @property string|null $modified_by
  * @property \CakeDC\Users\Model\Entity\User|null $modifier
  *
@@ -36,7 +36,7 @@ class ElectricityMeterReading extends Entity
      *
      * @var array<string, bool>
      */
-    protected $_accessible = [
+    protected array $_accessible = [
         'name' => true,
         'access_point_id' => true,
         'reading_date' => true,
