@@ -11,9 +11,9 @@ use Cake\ORM\Entity;
  * @property string $id
  * @property string|null $name
  * @property string|null $note
- * @property \Cake\I18n\FrozenTime|null $created
+ * @property \Cake\I18n\DateTime|null $created
  * @property string|null $created_by
- * @property \Cake\I18n\FrozenTime|null $modified
+ * @property \Cake\I18n\DateTime|null $modified
  * @property string|null $modified_by
  *
  * @property \App\Model\Entity\AppUser $creator
@@ -33,7 +33,7 @@ class PaymentPurpose extends Entity
      *
      * @var array<string, bool>
      */
-    protected $_accessible = [
+    protected array $_accessible = [
         'name' => true,
         'note' => true,
         'created' => true,
