@@ -18,10 +18,10 @@ use Cake\ORM\Entity;
  * @property string|null $serial_number
  * @property string|null $software_version
  * @property string|null $firmware_version
- * @property \Cake\I18n\FrozenTime|null $created
+ * @property \Cake\I18n\DateTime|null $created
  * @property string|null $created_by
  * @property \CakeDC\Users\Model\Entity\User|null $creator
- * @property \Cake\I18n\FrozenTime|null $modified
+ * @property \Cake\I18n\DateTime|null $modified
  * @property string|null $modified_by
  * @property \CakeDC\Users\Model\Entity\User|null $modifier
  * @property string|null $customer_connection_id
@@ -49,7 +49,7 @@ class RouterosDevice extends Entity
      *
      * @var array<string, bool>
      */
-    protected $_accessible = [
+    protected array $_accessible = [
         'name' => true,
         'access_point_id' => true,
         'device_type_id' => true,
