@@ -14,10 +14,10 @@ use Cake\ORM\Entity;
  * @property string|null $ip_address
  * @property string|null $ip_network
  * @property int|null $interface_index
- * @property \Cake\I18n\FrozenTime|null $created
+ * @property \Cake\I18n\DateTime|null $created
  * @property string|null $created_by
  * @property \CakeDC\Users\Model\Entity\User|null $creator
- * @property \Cake\I18n\FrozenTime|null $modified
+ * @property \Cake\I18n\DateTime|null $modified
  * @property string|null $modified_by
  * @property \CakeDC\Users\Model\Entity\User|null $modifier
  *
@@ -37,7 +37,7 @@ class RouterosDeviceIp extends Entity
      *
      * @var array<string, bool>
      */
-    protected $_accessible = [
+    protected array $_accessible = [
         'routeros_device_id' => true,
         'name' => true,
         'ip_address' => true,
