@@ -20,6 +20,9 @@ use AuditLog\Meta\RequestMetadata;
 use Cake\Cache\Cache;
 use Cake\Controller\Controller;
 use Cake\Core\Configure;
+use Cake\Datasource\Paging\PaginatedInterface;
+use Cake\Datasource\QueryInterface;
+use Cake\Datasource\RepositoryInterface;
 use Cake\Event\EventInterface;
 use Cake\Event\EventManager;
 use Cake\Http\Exception\NotFoundException;
@@ -52,7 +55,6 @@ class AppController extends Controller
     {
         parent::initialize();
 
-        $this->loadComponent('RequestHandler');
         $this->loadComponent('Flash');
 
         /*
@@ -65,13 +67,17 @@ class AppController extends Controller
     /**
      * Customize pagination
      *
-     * @param \Cake\ORM\Table|string|\Cake\ORM\Query|null $object Table to paginate
+     * @param \Cake\Datasource\RepositoryInterface|\Cake\Datasource\QueryInterface|string|null $object Table to paginate
      * (e.g: Table instance, 'TableName' or a Query object)
-     * @param array $settings The settings/configuration used for pagination.
-     * @return \Cake\ORM\ResultSet|\Cake\Datasource\ResultSetInterface|null Query results
+     * @param array<string, mixed> $settings The settings/configuration used for pagination. See {@link \Cake\Controller\Controller::$paginate}.
+     * @return \Cake\Datasource\Paging\PaginatedInterface
+     * @link https://book.cakephp.org/4/en/controllers.html#paginating-a-model
+     * @throws \Cake\Http\Exception\NotFoundException When a page out of bounds is requested.
      */
-    public function paginate($object = null, $settings = [])
-    {
+    public function paginate(
+        RepositoryInterface|QueryInterface|string|null $object = null,
+        array $settings = []
+    ): PaginatedInterface {
         try {
             // set maximal limit
             $this->paginate['maxLimit'] = 10000;
@@ -97,8 +103,10 @@ class AppController extends Controller
     /**
      * Global beforeFilter
      *
-     * @param \Cake\Event\EventInterface $event An Event instance
+     * @param \Cake\Event\EventInterface<\Cake\Controller\Controller> $event An Event instance
      * @return \Cake\Http\Response|null|void
+     * @link https://book.cakephp.org/4/en/controllers.html#request-life-cycle-callbacks
+     * @phpcsSuppress SlevomatCodingStandard.TypeHints.ReturnTypeHint.MissingNativeTypeHint
      */
     public function beforeFilter(EventInterface $event)
     {
@@ -181,7 +189,7 @@ class AppController extends Controller
      * @param array $query Query parameters to be added to the URL
      * @return string
      */
-    public function urlWithQuery(array $query = [])
+    public function urlWithQuery(array $query = []): string
     {
         $request = $this->getRequest();
 
